refactor: 상태를 가지지 않는 객체 제거해보기
