+++ README.md
@@ -31,3 +31,4 @@
 - [x] View 영역을 명확히 구분한다.
 - [x] 로또를 생성할 때 정적 팩토리 메서드를 활용해본다.
 - [x] 무분별한 getter를 줄이고 의미가 드러나도록 메서드 사용
+- [x] 상태를 가지지 않는 객체가 정말 필요한지 파악하고 변경
