+++ src/main/java/lotto/view/InputView.java
@@ -1,7 +1,6 @@
 package lotto.view;
 
 import lotto.domain.Lotto;
-import lotto.domain.Parser;
 
 import java.util.Scanner;
 
@@ -17,6 +16,6 @@ public class InputView {
 
     public static Lotto winningNumbers() {
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
-        return Parser.parseWinningNumbers(scanner.nextLine());
+        return Lotto.from(scanner.nextLine());
     }
 }
