+++ src/test/java/lotto/domain/LottosTest.java
@@ -11,8 +11,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class LottosTest {
     @Test
     void 당첨_개수() {
-        Lottos lottos = new Lottos(List.of(Lotto.from(List.of(1,2,3,4,5,6)), Lotto.from(List.of(4,5,6,7,8,9))));
-        assertThat(lottos.getMatchNums(Lotto.from(List.of(4,5,6,11,12,13)))).isEqualTo(Map.of(3, 2));
+        Lottos lottos = new Lottos(List.of(Lotto.from("1,2,3,4,5,6"), Lotto.from("4,5,6,7,8,9")));
+        assertThat(lottos.getMatchNums(Lotto.from("4,5,6,11,12,13"))).isEqualTo(Map.of(3, 2));
     }
 
     @Test
@@ -31,5 +31,12 @@ public class LottosTest {
     void 가격으로_생성_시_정렬_검증() {
         Lottos lottos = Lottos.from(2000);
         assertThat(lottos.isAllSorted()).isTrue();
+    }
+
+    @Test
+    void ROI_계산() {
+        Lottos lottos = new Lottos(List.of(Lotto.from("1,2,3,4,5,6")));
+        Map<Integer, Integer> matchNums = lottos.getMatchNums(Lotto.from("1,2,3,11,12,13"));
+        assertThat(lottos.calculateROI(matchNums)).isEqualTo(5f);
     }
 }
